spec: added a spec for napkin
I've written a spec/plan that outlines clearly the function and scope of
napkin so that the project remains small, completable, concise and
opinionated; it has a task and it does it in a dead-simple way without
fluff and unnecessary features; bloat-free software.

Some specs on napkin are also written down so that the napkin home
directory is compatible between napkin rewrites and versions without
ambiguity.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>