spec changes: specified more functionality
I've changed the formatting of the specification to use 72 character
lines instead of 80 character lines to make it more readable.

I've also renamed the original export command to the 'backup' command
as that's a more accurate description of the purpose of the command.

There is also now an `export` command that's been specified that
*actually* exports the napkins as human readable text files that cannot
be re-imported due to the loss of meta-data.

Behaviour for invalid yaml files have also been specified so that the
program doesn't just crash and can report errors to the user without
having to use stdout as that would go against the design and purpose
of using temporary files for input.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>