interactive yaml configs with robust error handling
I've fully implemented the basic functions needed for napkin to use yaml
configurations with basic error reporting through repoening the config
file with the error as a comment on the afflicted line (as shown in the
spec).

I've only implemented the most basic functions as any other
functionality required with parsing the yaml (and subsequent error
handling) can be done with a combination of the functions provided.

A quick demo of how the configuration changes should look has also been
written in the main function of the program.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>