cli: added version flag, help message & rm command
I've added a simple help message for the cli based off of the spec that
I'd written earlier.

I've also added a version flag so it's trivial to figure out which
version of napkin is being run.

I've also removed the command parsing function from the cli as it's
pretty redundant and inefficient to compare against the fields of an
enum and then switch on the enum when you can just do it in one if
statement without the enum entirely.

I've also made the runCli function return an error at the end if no
valid command or option is found.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>