context: added context command and context editing
I've implemented the context command for editing the context file with
temporary-files and a user-provided editor.

Similar to the rust version, it reports errors by annotating the string
that the user edits with an error message, though, unlike the rust
version, it has smart memory management so that everytime the user makes
a change, it reallocates and frees on it's own.

Error metadata has also been removed as it's outside of the v1 scope of
this project.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>