locks: actually implemented file-based locks
I've actually implemented the file-based locks instead of the
file-system FLOCK based locks as I don't trust them.

The second reason for this, is that file-based locks are easy and more
ergonomic to implement (with defer) than the file-system based locks
which aren't stable.

Also added a quick 'pathExists' function to make my life easier.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>