added napkins creation, new command & better dumps
I've added support for timestamps by added another dependency for it.

I've added the ability to update the context.yml file with new napkins
automatically without need for users to do it.

I've also added functionality that allows for new napkins to be
initialised alongside it's cli command all according to the spec.

The extensions for the napkins are grabed from splitting the napkin
based on '.' as that's the easiest way I know how.

There is also graceful user dumping handling so that there isn't a scary
backtrace when the user crashes the program on purpose.

Locking may look a bit weird due to the need to protect on-going user
edits and the conflicts between functions that also lock.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>