meta command: added support for editing metadata
I've added the meta-data command according to the spec.

There's also basic cli input implemented for this command, I've done it
this way as it's the most simple and efficient.

Richer errors would be nice, though that's out of the scope of this
project.

I've made the metaPath function to make it the action reusable and
more ergonomic to write.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>