version checker, latestContents, edit cmd & bugfix
This is a very large commit.

The first change made in this commit was a relatively simple version
checker that also initialises the napkin home if it doesn't exist
already. This is done so that corruption cannot happen between versions
of napkin and to also bundle together common functionality. Note that
editing the context file doesn't check for the version, as that
functionality should be version agnostic as if you cannot even edit the
context file then you can't extract the version from it either.

Some other minor changes were some bugfixes with the locks so that they
unlocked even on edge-cases. And also creating a new napkin would
actually return the id of the napkin that's been initialised.

The largest change, or feature, was the 'edit' command of the cli, it
was implemented according to the spec and implementation should make
sense. However, the actual editing of the metadata-file of the napkin
is done through string manipulation instead of the more obvious approach
of parsing it, modifying the yaml document and then turning it into a
string. This was chosen against, due to a weird arraylisthashmap error
that'd occur once there were enough napkins edits in the history, and
only when modifying the document through the 'put' function.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>