added better file-extension support in napkin meta
I've changed the spec to have better support for changing the file
extension for individual edits in the history of a napkin as previously
that action would just corrupt the napkin database.

Do note though, after an edit is made, the file extension cannot be
modified without making that edit corrupt or lost. You can only modify
the top-level filename's extension for future edits. This is due to
implementation details where the underlying text file is stored with the
timestamp with that specified extension, so that while editing, you can
have syntax highlighting.

I've also removed the filename attribute of napkins in favour of a split
filename with the name, and file extension attributes which would come
together to determine the final exported filename. This is more
favourable as it gives the name attribute an actual purpose that the
description coudln't fulfill and it also removed any abmiguity with file
extensions.

Also the gist was also added to the spec.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>