major sweeping ergonomic UX changes
I've completely redesigned some aspects of napkin to make it as
pain-free and easy to write down back-of-the-napkin ideas (the whole
point of this project) as previously, you had to create a new napkin,
write down it's id, then edit it while modifying a bunch of metadata
before you could even write down your idea which is really annoying and
defeats the purpose of using this cli instead of just a simple git repo.

The first major change was the use of human readable string-ids instead
of timestamps. The reasoning behind this should be self-explainitory,
it's much easier to write 'napkin edit hello-world' instead of
'napkin edit 10091283057361'. Plus there were no benefits from timestamp
ids for whole napkins as the only real difference is the ability to have
two napkins with the same name, which would've been confusing anyways.

The second major change came to napkin's 'new' command, it's been
changed so that you're able to write down the idea first before worrying
about metadata or descriptions, as you can't describe something before
you write it. It also makes writing down ideas trivial and pain-free
with little friction.

I've also written a bunch of optimisations in which certain safety
checks and locks are removed as those functions aren't intended to be
used standalone, so it shifts the responsiblity of ensuring the napkin
home existing to the caller rather than the callee.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>