export: added support for the exporting of napkins
I've added support for exporting napkins into a specified directory
as according to the spec.

Note that the weird double check of if the output path exists is due to
having to fix FileNotFound errors while creating the dirs due to having
to deal with both relative and absolute paths.

This should be the last core feature of napkin (dis-regarding backups
as that is a large undertaking due to the requirement of tarballs).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>