backup command: added the backup command
I've quickly added a simple backup command (mainly through copying the
export command and modifying it slightly).

The backup command now backs up to an output directory instead of a
tarball as I wasn't able to get that working without errors so far.

The backup command also only copies the referenced files, so it sort of
auto-cleans.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>